basic spawning (#21)
* moving goblin_torch to /scenes/mob

* Skeleton asset added

"32rogues" may be used in both non-commercial and commercial projects, excluding those relating to or containing non-fungible tokens ("NFT"),  blockchain-related projects. It may not be used in conjunction with generative artificial intelligence proejcts or machine learning projects. You may modify it to suit your needs. You may not redistribute or resell it. Credit is not necessary, but very much appreciated.

Copyright 2024 Seth Boyles

* Skeleton archer asset

"32rogues" may be used in both non-commercial and commercial projects, excluding those relating to or containing non-fungible tokens ("NFT"),  blockchain-related projects. It may not be used in conjunction with generative artificial intelligence proejcts or machine learning projects. You may modify it to suit your needs. You may not redistribute or resell it. Credit is not necessary, but very much appreciated.

Copyright 2024 Seth Boyles

* Assets added

same license as earlier

* added AI node to goblin torch

* goblin spawn logic

it spawns but it's not connected to NPC.gd or AI.gd

* Update basic_melee.tscn